+++ packages/carbon-graphs/dev/examples/controls/gantt.js
@@ -615,7 +615,6 @@ export const renderGanttPanning = (id) => {
   });
   return graph;
 };
-
 export const renderGanttPanningWithDynamicData = (id) => {
   const axisData = utils.deepClone(getDemoData(`#${id}`, 'GANTT'));
   axisData.showActionLegend = true;
@@ -668,7 +667,57 @@ export const renderGanttPanningWithDynamicData = (id) => {
   });
   return graph;
 };
+export const renderGanttPanningWithDynamicEventline = (id) => {
+  const axisData = utils.deepClone(getDemoData(`#${id}`, "GANTT"));
+  axisData.showActionLegend = true;
+  axisData.axis.x.lowerLimit = new Date(2016, 0, 1, 0).toISOString();
+  axisData.axis.x.upperLimit = new Date(2016, 0, 2, 0).toISOString();
+  axisData.eventline = [
+    {
+      color: Carbon.helpers.COLORS.GREY,
+      style: {
+        strokeDashArray: "4,4"
+      },
+      value: new Date(2016, 0, 1, 4).toISOString()
+    }
+  ];
+  axisData.pan = {
+    enabled: true
+  };
+  const graphData = {
+    key: "track 0",
+    trackLabel: {
+      display: "Default",
+      onClick: loadXAndYAxisLabelPopup
+    },
+    tasks: tasks[5],
+    actions: actions[2],
+    events: events[1],
+    activities: activities[3]
+  };
+  const createGraph = () => {
+    graphData.eventline = [
+      {
+        color: Carbon.helpers.COLORS.BLACK,
+        style: {
+          strokeDashArray: "2,2"
+        },
+        value: new Date(2016, 0, 1, 6, 30).toISOString()
+      }
+    ];
+    graph.reflow(graphData);
+  };
 
+  const graph = Carbon.api.gantt(axisData);
+  graph.loadContent(graphData);
+  axisData.axis = graph.config.axis;
+
+  createPanningControls(id, {
+    axisData,
+    creationHandler: createGraph
+  });
+  return graph;
+};
 export const renderGanttAction = (id) => {
   const data = utils.deepClone(getDemoData(`#${id}`, 'GANTT'));
   data.dateline = [];