+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## Unreleased
 
+* Changed
+  * Updated reflow in Graph and Gantt construct to update eventline.
+
 -   Migrated to the terra-graphs mono-rep.